рефактор кода: uni. компоненты-обертки
