+++ src/App.jsx
@@ -1,11 +1,9 @@
 import './App.css';
-import { Container } from './components/mainWrappers_components/container';
+import AppRender from './pages/appRender';
 
 function App() {
   return (
-    <div className="wrapper">
-     <Container/>
-  </div>
+   <AppRender/>
   )
 }
