+++ src/components/basicPage/barContentForm.jsx
@@ -1,15 +1,18 @@
 import BarPlayer from './barPlayer'
-import VolumeBlock from '../volumeBlock_component/volumeBlock'
+import VolumeBlock from '../volumeBlock/volumeBlock'
 
 const BarContentForm = () => {
   return (
-    <div className="bar__content">
+    <div className="bar">
+  <div className="bar__content">
     <div className="bar__player-progress"></div>
     <div className="bar__player-block">
       <BarPlayer />
       <VolumeBlock />
     </div>
   </div>
+    </div>
+  
   )
 }
