+++ src/components/mainWrappers/footer.jsx
@@ -1,0 +1,8 @@
+export function Footer() {
+    return (
+        <footer className="footer"></footer>
+    );
+  }
+  
+
+  export default Footer
