+++ src/components/player/playerControls.jsx
