--- src/pages/appMainPage.jsx
@@ -1,8 +1,0 @@
-import Wrapper from "../components/mainWrappers_components/wrapper";
-
-function AppRender() {
-    <Wrapper />
-  }
-  
-  export default AppRender;
-
